organized project structure into ./src
