+++ gui_components/edit/commands.py
@@ -1,7 +1,7 @@
 import tkinter as tk
 
-import config
-from components import Point
+from src.common import config
+from src.routine.components import Point
 from gui_components.interfaces import Frame
