+++ src/routine/routine.py
@@ -1,12 +1,10 @@
 """A collection of classes used in the 'machine code' generated by Auto Maple's compiler for each routine."""
 
-import config
-import utils
+from src.common import config, settings, utils
 import csv
-import settings
 from os.path import splitext, basename
-from components import Point, Label, Jump, Setting, Command, SYMBOLS
-from layout import Layout
+from src.routine.components import Point, Label, Jump, Setting, Command, SYMBOLS
+from src.routine.layout import Layout
 
 
 def update(func):
